+++ sqlspec/utils/serializers/_schema.py
@@ -125,14 +125,18 @@ def _dump_identity_dict(value: Any) -> "dict[str, Any]":
 
 
 def _dump_msgspec_fields(value: Any) -> "dict[str, Any]":
-    return {field_name: value.__getattribute__(field_name) for field_name in value.__struct_fields__}
+    from msgspec import structs
+
+    return {field.encode_name: value.__getattribute__(field.name) for field in structs.fields(type(value))}
 
 
 def _dump_msgspec_excluding_unset(value: Any) -> "dict[str, Any]":
+    from msgspec import structs
+
     return {
-        field_name: field_value
-        for field_name in value.__struct_fields__
-        if (field_value := value.__getattribute__(field_name)) != UNSET
+        field.encode_name: field_value
+        for field in structs.fields(type(value))
+        if (field_value := value.__getattribute__(field.name)) != UNSET
     }